Move DATETIME_FORMAT from app to app.utils
To avoid cyclical import issues
