+++ app/__init__.py
@@ -31,10 +31,6 @@ from app.clients.sms.firetext import FiretextClient
 from app.clients.sms.mmg import MMGClient
 from app.clients.performance_platform.performance_platform_client import PerformancePlatformClient
 
-DATETIME_FORMAT_NO_TIMEZONE = "%Y-%m-%d %H:%M:%S.%f"
-DATETIME_FORMAT = "%Y-%m-%dT%H:%M:%S.%fZ"
-DATE_FORMAT = "%Y-%m-%d"
-
 
 class SQLAlchemy(_SQLAlchemy):
     """We need to subclass SQLAlchemy in order to override create_engine options"""
