+++ app/utils.py
@@ -11,8 +11,10 @@ from notifications_utils.template import (
     BroadcastMessageTemplate,
 )
 
-from app import DATETIME_FORMAT
 
+DATETIME_FORMAT_NO_TIMEZONE = "%Y-%m-%d %H:%M:%S.%f"
+DATETIME_FORMAT = "%Y-%m-%dT%H:%M:%S.%fZ"
+DATE_FORMAT = "%Y-%m-%d"
 local_timezone = pytz.timezone("Europe/London")
