+++ tests/app/v2/template/test_get_template.py
@@ -2,8 +2,8 @@ import pytest
 
 from flask import json
 
-from app import DATETIME_FORMAT
 from app.models import (TEMPLATE_TYPES, EMAIL_TYPE, SMS_TYPE, LETTER_TYPE,)
+from app.utils import DATETIME_FORMAT
 from tests import create_authorization_header
 from tests.app.db import create_template, create_letter_contact
